Bug fixes, improvements, and a new command-line argument.
The JSBML version has been updated to revision 2492.

It is now possible to disable the use of URIs for the annotation of
models that are not (yet) included in the MIRIAM registry using the new
command-line argument --accept-only-miriam-uris=<true|false>.

Several improvements were made in the COBRA parser: a source of an
IndexOutOfBoundsExceptions was solved, error messages have been
improved, special cases of common errors in mat files were integrated
and are now caught rather than causing failure.

A typo was fixed in the model notes file.