Final changes for 1.6 release
Added licensing information for all dependencies, as far as possible.
Licenses for registry-lib and sysbio-1390 are still missing - have to figure out how to make mvn recognize the pom files.
Added steps to pom.xml to pack licenses into both the zip and the jar when building the project.
Build command for maven is now back to mvn clean compile package.